+++ .gitattributes
@@ -1,1 +1,1 @@
-*.bin filter=lfs diff=lfs merge=lfs -text
+wiki_tokened.bin filter=lfs diff=lfs merge=lfs -text
